# Conflicts:
#	.eslintcache
#	src/main/java/org/lamisplus/modules/ndr/installers/SchemaInstaller3.java
#	src/main/java/org/lamisplus/modules/ndr/service/NDRJSONService.java
#	src/main/java/org/lamisplus/modules/ndr/service/NDRService.java
#	src/main/resources/views/static/ndr/asset-manifest.json
#	src/main/resources/views/static/ndr/index.html

+++ src/main/java/org/lamisplus/modules/ndr/domain/dto/NdrXmlStatusDto.java
@@ -17,6 +17,6 @@ public class NdrXmlStatusDto implements Serializable {
     private  String facility;
     private  LocalDateTime lastModified;
     private String pushIdentifier;
-    private int percentagePushed;
+    private Long percentagePushed;
     private Boolean completelyPushed;
 }
